타입수정

+++ x만큼-간격이-있는n-개의-숫자/java/20220901/src/Solution.java
@@ -1,9 +1,9 @@
 class Solution {
-    public long[] solution(int x, int n) {
+    public long[] solution(long x, int n) {
         long[] answer = new long[n];
 
-        for (long i = 0; i < n; i += 1) {
-            answer[(int) i] = x * (i + 1);
+        for (int i = 0; i < n; i += 1) {
+            answer[i] = x * (i + 1);
         }
 
         return answer;
